async_ai_bot/requirements.txt
fastapi==0.112.0
httpx==0.27.0
openai==1.40.3
pytest==8.3.2
python-dotenv==1.0.1
uvicorn[standart]
websockets==12.0